+++ server/server.go
@@ -11,11 +11,9 @@ import (
 	"net"
 	"net/http"
 	"os"
-	"os/signal"
 	"runtime"
 	"strconv"
 	"sync"
-	"syscall"
 	"time"
 
 	// Allow dynamic profiling.
@@ -180,30 +178,6 @@ func PrintAndDie(msg string) {
 func PrintServerAndExit() {
 	fmt.Printf("nats-server version %s\n", VERSION)
 	os.Exit(0)
-}
-
-// Signal Handling
-func (s *Server) handleSignals() {
-	if s.opts.NoSigs {
-		return
-	}
-	c := make(chan os.Signal, 1)
-
-	signal.Notify(c, syscall.SIGINT, syscall.SIGUSR1)
-
-	go func() {
-		for sig := range c {
-			Debugf("Trapped %q signal", sig)
-			switch sig {
-			case syscall.SIGINT:
-				Noticef("Server Exiting..")
-				os.Exit(0)
-			case syscall.SIGUSR1:
-				// File log re-open for rotating file logs.
-				s.ReOpenLogFile()
-			}
-		}
-	}()
 }
 
 // Protected check on running state
